api/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP Client
httpx==0.26.0

# Date/Time
python-dateutil==2.8.2

# CORS
fastapi-cors==0.0.6

# Utilities
python-slugify==8.0.1

# AI APIs (already in use)
anthropic==0.39.0